requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
fastapi==0.115.12
fire==0.7.0
fonttools==4.58.2
fpdf==1.7.2
h11==0.16.0
idna==3.10
img2pdf==0.6.1
lxml==5.4.0
numpy==2.2.6
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2docx==0.5.8
pikepdf==9.8.1
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-docx==1.1.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.2
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
wrapt==1.17.2
